requirements/base.txt
uWSGI >=2.0, <2.1
psycopg2
Django >=1.7, <1.8
djangorestframework >=3.1, <3.2
djangorestframework-jwt
django-filter
markdown
pytz
